+++ src/Components/LabelTextInput.css
@@ -1,4 +1,5 @@
 .labeltextinput_container {
+  display: flex;
   flex-direction: column;
   margin: 10px;
 }
@@ -12,6 +13,6 @@
 
 .labeltextinput_input {
   min-height: 50px;
-  width: 100%;
+
   display: block;
 }
